Directional Shooting
Added in directional shooting: the player can now shoot in whatever
direction they're pressing with the arrow keys. Additionally, there is
now a gun on top of the player ship to help the player visualize this.

Also, there is a new method in weapon that allows you to manually set
where the bullets come from via code. This was used to make the shots
come out of the player's rotating gun.