Predictive Shots
Wrote a script that will shoot projectiles at where a target will be if
they do not change course.

Note: If the bullets are not fast enough to aim in front of the target,
they will simply target the target's current position. Make sure the
bullets are fast enough to aim ahead.