Changing Icon + Flash when hit
Weapon Pickups now have an Icon associated with their weapon in the UI.
When you pick up a weapon the UI ammo counter will show which weapon it
has (using some icons I threw together in paint, they can and probably
should be changed for the final release.)

The screen flashes red for a very brief period when hit. We should
consider making this an optional setting, as it might be tough on the
eyes for some people?
Also, when the player dies the redness stays. I can easily fix it, but I
personally find it pretty cool.